Time: 1255 ms (5.01%) | Memory: 41.8 MB (87.32%) - LeetSync

+++ 367-valid-perfect-square/valid-perfect-square.java
@@ -3,7 +3,7 @@ class Solution {
         if(num==1){
             return true;
         }
-        for(long i=1;i<num;i++){
+        for(int i=1;i<num;i++){
             if(i*i==num){
                 return true;
             }
